refactor(design): extract cache and merge helpers
